+++ packages/client/src/api/getVenues.ts
@@ -2,6 +2,6 @@ import { api } from "../utils/generateApi";
 
 export const venues = api({
   methodName: "venues",
-  path: "/v2/fungible/venues",
+  path: "v2/fungible/venues",
   transformResponse: (response) => response.venues,
 });