+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # nextjs
 
+## 0.1.109
+
+### Patch Changes
+
+- @skip-go/widget@3.10.11
+
+## 0.1.108
+
+### Patch Changes
+
+- @skip-go/widget@3.10.10
+
 ## 0.1.107
 
 ### Patch Changes